티켓 예매 작업 진행중 - 티켓 결제 후 캐시 차감 - 캐시 충전 - 티켓 등록 오류 수정

+++ app/src/main/java/com/ezen/lolketing/view/main/shop/PurchaseActivity.java
@@ -50,7 +50,7 @@ public class PurchaseActivity extends AppCompatActivity {
 
     private static final int SEARCH_ADDRESS_ACTIVITY = 10000;
     private String id;
-    int cache = 0;
+    Long cache = 0L;
     String phone;
     String nickname;
